+++ website/src/entry/browser-warning/browser-warning.scss
