+++ website/src/index.html
@@ -25,5 +25,7 @@
   <body>
     <div id="app"></div>
     <noscript>Please enable JavaScript to use ${htmlWebpackPlugin.options.brandName}</noscript>
+
+  <script src="/${htmlWebpackPlugin.options.browserWarningPath}" async defer></script>
   </body>
 </html>
