+++ website/webpack/webpack.config.common.js
@@ -33,10 +33,7 @@ const commonConfig = merge([
     // convenient with more complex configurations.
     entry: {
       // This will build an app.js file from the `main` module.
-      app: ['main'],
-      // Browser warning, exported separately so it doesn't have anything that IE11
-      // does not support
-      browser: ['bootstrapping/browser'],
+      app: ['entry/main'],
     },
     context: parts.PATHS.src,
     output: {
